Make `EnclaveAttestationProvider` public (#3626)
