+++ src/node/rpc/member_frontend.h
@@ -7,13 +7,13 @@
 #include "ccf/crypto/key_pair.h"
 #include "ccf/ds/nonstd.h"
 #include "ccf/json_handler.h"
+#include "ccf/node/quote.h"
 #include "ccf/service/tables/gov.h"
 #include "ccf/service/tables/jwt.h"
 #include "ccf/service/tables/members.h"
 #include "ccf/service/tables/nodes.h"
 #include "frontend.h"
 #include "js/wrap.h"
-#include "node/quote.h"
 #include "node/rpc/call_types.h"
 #include "node/rpc/gov_effects_interface.h"
 #include "node/rpc/node_operation_interface.h"
