+++ src/node/rpc/node_operation_interface.h
@@ -2,6 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/node/quote.h"
 #include "ccf/node_startup_state.h"
 #include "ccf/node_subsystem_interface.h"
 #include "ccf/quote_info.h"
@@ -11,14 +12,6 @@
 
 namespace ccf
 {
-  enum class QuoteVerificationResult
-  {
-    Verified = 0,
-    Failed,
-    FailedCodeIdNotFound,
-    FailedInvalidQuotedPublicKey,
-  };
-
   using ExtendedState = std::tuple<
     NodeStartupState,
     std::optional<kv::Version> /* recovery_target_seqno */,
